Refactor SpellBooks component to receive spellBooks and onAddSpellBook as props
